Separating out news task
