+++ tasks/fmOptions.json
@@ -1,0 +1,4 @@
+{
+  "property": "frontMatter",
+  "remove": true
+}
