+++ tasks/renamePage.js
@@ -1,0 +1,13 @@
+var path = require("path");
+
+function renamePage(filePath) {
+  if (filePath.basename !== "index") {
+    filePath.dirname = path.join(
+      filePath.dirname,
+      filePath.basename
+    );
+    filePath.basename = "index";
+  }
+}
+
+module.exports = renamePage;
